AOSCOS: Fix build error '__max_packages' undeclared on non-SMP platforms
The commit "BACKPORT: DEEPIN: LoongArch: Adjust the calculation of the
number of packages", introduced a build failure on non-SMP configurations
due to the unconditional use of '__max_packages', which is only defined
when CONFIG_SMP is enabled.

  arch/loongarch/kernel/setup.c: In function 'parse_cpu_table':
  arch/loongarch/kernel/setup.c:150:9: error: '__max_packages' undeclared (first use in this function)
    150 |         __max_packages++;
        |         ^~~~~~~~~~~~~~
  arch/loongarch/kernel/setup.c:150:9: note: each undeclared identifier is reported only once for each function it appears in

Fix this by guarding the '__max_packages' increment with CONFIG_SMP,
as the variable is only relevant for SMP systems.

Link: https://github.com/AOSC-Tracking/linux/commit/e48a6614dda6958e08efa37c35364ed2429c60d6

Signed-off-by: Ilikara <3435193369@qq.com>