RISCV64: SG2042: HACK: ttm: disallow cached mapping
FIXME: this hack fixed the issue when booting PCIe radeon GPU card,
it will report:
[    2.909876] [drm:r600_ring_test] *ERROR* radeon: ring 0 test failed (scratch(0x8504)=0xCAFEDEAD)

Rootcasue Analysis:
The radeon/amdgpu driver expects pcie to snoop cpu cache, which
means the contents of cpu cache should also be seen by pcie.
But sg2042 does not meet this requirement. When the GPU is being
initialized, the GPU's ring_test tries to write a value to a
certain memory address, and then read back to check whether it
is written in (before that, the cpu writes another value in).
Due to SG2042's issue, gpu cannot snoop cpu cache, check fails.

The hack solution is to change the memory mapped by the ttm
memory manager to uncached mode to workaround this SoC issue.

Signed-off-by: Icenowy Zheng <uwu@icenowy.me>
Tested-by: Chen Wang <unicorn_wang@outlook.com>
Signed-off-by: Han Gao <rabenda.cn@gmail.com>