rust: device: Move property_present() to FwNode
The new FwNode abstraction will be used for accessing all device
properties.

It would be possible to duplicate the methods on the device itself, but
since some of the methods on Device would have different type sigatures
as the ones on FwNode, this would only lead to inconsistency and
confusion. For this reason, property_present is removed from Device and
existing users are updated.

Signed-off-by: Remo Senekowitsch <remo@buenzli.dev>
Acked-by: Viresh Kumar <viresh.kumar@linaro.org>