AOSCOS: gpio: loongson-64bit: Add LS7A GPIO interrupt support
On IPASON NL38-N11 laptops, the touchpad device (a HID multitouch device
connected via I2C) uses GPIO 50 to send interrupts. Unfortunately,
gpio-loongson-64bit does not support interrupt for LS7A yet, causing the
touchpad fail to be probed with:

    i2c_hid_acpi i2c-PNP0C50:00: HID over i2c has not been provided an Int IRQ

Add the offset of the GPIO_INT_EN controlling register to
loongson_gpio_ls7a_data to support the interrupt.

And, in LS7A the GPIO 4 ... 56 share the same interrupt pin, we need to
take this fact into account when calling platform_get_irq or we'd fail
with "IRQ index 50 not found."

Signed-off-by: Xi Ruoyao <xry111@xry111.site>

Signed-off-by: Kexy Biscuit <kexybiscuit@aosc.io>